feat(shimmer): move every shimmer to its own file
